feat: refactor home
