+++ src/components/home/defaultColumns.ts
@@ -1,0 +1,94 @@
+const defaultColumns = [
+  {
+    id: 1,
+    name: 'Rank',
+    content: 'Rank',
+    value: 'rank',
+    sortable: true,
+    isOptional: false,
+    isSelectedAtStart: true,
+  },
+  {
+    id: 2,
+    name: 'Market',
+    content: '',
+    value: 'market',
+    sortable: false,
+    isOptional: false,
+    isSelectedAtStart: true,
+  },
+  {
+    id: 3,
+    name: 'Name',
+    content: 'Name',
+    value: 'name',
+    sortable: true,
+    isOptional: false,
+    isSelectedAtStart: true,
+  },
+  {
+    id: 4,
+    name: 'Price',
+    content: 'Price',
+    value: 'price',
+    sortable: true,
+    isOptional: false,
+    isSelectedAtStart: true,
+  },
+  {
+    id: 5,
+    name: '24H Change',
+    content: '24H Change',
+    value: 'change',
+    sortable: true,
+    isOptional: true,
+    isSelectedAtStart: false,
+  },
+  {
+    id: 6,
+    name: 'Deposits',
+    content: 'Deposits',
+    value: 'deposits',
+    sortable: true,
+    isOptional: true,
+    isSelectedAtStart: true,
+  },
+  {
+    id: 7,
+    name: '% Locked',
+    content: '% Locked',
+    value: 'locked',
+    sortable: true,
+    isOptional: true,
+    isSelectedAtStart: true,
+  },
+  {
+    id: 8,
+    name: '1YR Income',
+    content: '1YR Income',
+    value: 'income',
+    sortable: true,
+    isOptional: true,
+    isSelectedAtStart: true,
+  },
+  {
+    id: 9,
+    name: 'Claimable Income',
+    content: 'Claimable Income',
+    value: 'claimable',
+    sortable: false,
+    isOptional: true,
+    isSelectedAtStart: false,
+  },
+  {
+    id: 10,
+    name: 'Trade',
+    content: '',
+    value: 'trade',
+    sortable: false,
+    isOptional: false,
+    isSelectedAtStart: true,
+  },
+]
+
+export default defaultColumns
