+++ src/components/index.tsx
@@ -29,7 +29,7 @@ export { default as FooterEmailForm } from './newsletter/FooterEmailForm'
 export { default as Tooltip } from './tooltip/Tooltip'
 export { default as PromoVideoModal } from './video/PromoVideoModal'
 export { default as AddToMetamaskButton } from './wallet/AddToMetamaskButton'
+export { default as HomeHeader } from './home/HomeHeader'
 export { default as A } from './A'
-export { default as Home } from './Home'
 export * from './mutual-tokens'
 export * from './layouts'
